Move docs left nav into component
